UC1 LED latency: dedup brightness + drain whole burst per worker tick
Two compounding bottlenecks made knob-ring updates feel laggy:

1. Gradient-mode pots re-emitted ALL ring cells on every push,
   even when nothing on a given cell had changed (cache only stored
   selection state, not brightness). Packed the cache to
   `uint16_t = sel | (brightness << 8)` so dedup catches both
   banks. A 22-frame Gradient burst now collapses to 2-4 frames
   per knob tick once the dot has settled into its move pattern.

2. UC1Device::workerLoop_ pulled ONE frame per iteration with a
   1 ms libusb event pump between each — a 22-frame burst stalled
   ~50 ms before the last frame went out. Now the loop drains the
   entire pending queue in one pass under the lock, then bursts
   everything back-to-back to the bulk OUT endpoint. Keepalive +
   GR streaming still fire on their own cadence after the burst.

+++ extension/src/UC1Surface.h
@@ -162,12 +162,12 @@ private:
     // when on, knob clicks move the param by 1/4 the normal amount.
     std::atomic<bool> fineMode_{false};
 
-    // Per-knob LED-ring cell state cache. pushKnobRing_ writes only
-    // cells whose target state differs from the cache; setFocusedTrack
-    // clears it so the next refresh re-writes every cell, preventing
-    // device/cache desync (init-flood lit cells we never told the
-    // firmware to clear, etc.).
-    std::unordered_map<uint8_t, std::vector<uint8_t>> ringCellCache_;
+    // Per-knob LED-ring cell state cache. Each entry packs the last-
+    // sent (selection_state | brightness_state << 8) so dedup catches
+    // brightness changes too, not just selection. pushKnobRing_ skips
+    // cells whose packed target matches the cache; setFocusedTrack /
+    // invalidateCache clears so the next refresh re-writes everything.
+    std::unordered_map<uint8_t, std::vector<uint16_t>> ringCellCache_;
 };
 
 } // namespace uc1
